[Spring] Rename Cukes to Cucumber
