+++ spring/src/test/java/cucumber/runtime/java/spring/dirtiescontextconfig/RunCucumberTest.java
@@ -1,12 +1,13 @@
 package cucumber.runtime.java.spring.dirtiescontextconfig;
 
 import cucumber.api.CucumberOptions;
-
 import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(glue = {"cucumber.runtime.java.spring.dirtiescontextconfig"},
-        features = {"classpath:cucumber/runtime/java/spring/dirtyCukes.feature"})
-public class RunCukesTest {
+@CucumberOptions(
+    glue = {"cucumber.runtime.java.spring.dirtiescontextconfig"},
+    features = {"classpath:cucumber/runtime/java/spring/dirtyCukes.feature"}
+)
+public class RunCucumberTest {
 }
