+++ spring/src/test/java/cucumber/runtime/java/spring/metaconfig/dirties/RunCucumberTest.java
@@ -1,0 +1,13 @@
+package cucumber.runtime.java.spring.metaconfig.dirties;
+
+import cucumber.api.CucumberOptions;
+import cucumber.api.junit.Cucumber;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(
+    glue = {"cucumber.runtime.java.spring.metaconfig.dirties"},
+    features = {"classpath:cucumber/runtime/java/spring/dirtyCukesWithMetaConfiguration.feature"}
+)
+public class RunCucumberTest {
+}
