+++ spring/src/test/java/cucumber/runtime/java/spring/metaconfig/general/RunCucumberTest.java
@@ -1,4 +1,4 @@
-package cucumber.runtime.java.spring.contextconfig;
+package cucumber.runtime.java.spring.metaconfig.general;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
@@ -7,14 +7,13 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
     glue = {
-        "cucumber.runtime.java.spring.contextconfig",
+        "cucumber.runtime.java.spring.metaconfig.general",
         "cucumber.runtime.java.spring.commonglue",
         "cucumber.api.spring"
     },
     features = {
-        "classpath:cucumber/runtime/java/spring/stepdefInjection.feature",
-        "classpath:cucumber/runtime/java/spring/transaction.feature"
+        "classpath:cucumber/runtime/java/spring/springBeanInjectionWithMetaConfiguration.feature"
     }
 )
-public class RunCukesTest {
+public class RunCucumberTest {
 }
